Make isint() 64-bit and stop timestring() returning empty on overflow
Chasing where "#-1 ARGUMENT MUST BE INTEGER" in the pose tracker actually came
from. Not sub() or div() — both are full Int64, sub(9223372036854775807,1) is
exact. The #-1 from msecs() propagated: sub() answered "ARGUMENTS MUST BE
NUMBERS", div() "ARGUMENTS MUST BE INTEGERS", and timestring() rejected that with
the singular message that reached the screen. The root cause was msecs(), as
fixed; the last link was timestring, not sub.

Two real defects turned up on the way.

isint() parsed with int.TryParse, so it answered 0 for every value above
2147483647 — the millisecond timestamps this very package stores among them —
while add(), sub() and div() all treated the same value as an integer. fun_isint
reads with parse_ival_full, which is parse_int64 (src/funmath.c:82-84,
1408-1416), so this was wrong on its own terms as well as inconsistent.

timestring() returned an EMPTY string for anything past ~9.2e11 seconds:
TimeSpan.FromSeconds overflows there, and empty is the one answer a caller cannot
distinguish from a real one. It now divides the seconds out directly, which
cannot throw anywhere in the non-negative long range.

PennMUSH refuses that input instead, its seconds being an "unsigned int"
(src/funtime.c:484-501). That ceiling is a property of the C type rather than of
timestring, and this server is 64-bit throughout, so it is deliberately not
reproduced — timestring(4294967296) formats here and errors there.

+++ SharpMUSH.Implementation/Functions/UtilityFunctions.cs
@@ -735,9 +735,15 @@ public partial class Functions
 		return new CallState(!(await Mediator!.Send(new GetObjectNodeQuery(parsed.AsValue()))).IsNone);
 	}
 
+	/// <summary>
+	/// <c>isint()</c> is 64-bit: <c>fun_isint</c> reads with <c>parse_ival_full</c>, which is
+	/// <c>parse_int64</c> (<c>src/funmath.c:82-84, 1408-1416</c>). Parsing as a 32-bit int said no to
+	/// every value above 2147483647 — a millisecond timestamp among them — while <c>add()</c>,
+	/// <c>sub()</c> and <c>div()</c> all handled the same value as an integer quite happily.
+	/// </summary>
 	[SharpFunction(Name = "isint", MinArgs = 1, MaxArgs = 1, Flags = FunctionFlags.Regular | FunctionFlags.StripAnsi)]
 	public static ValueTask<CallState> IsInt(IMUSHCodeParser parser, SharpFunctionAttribute _2) =>
-		ValueTask.FromResult<CallState>(new(int.TryParse(parser.CurrentState.Arguments["0"].Message!.ToString(), out var _) ? "1" : "0"));
+		ValueTask.FromResult<CallState>(new(long.TryParse(parser.CurrentState.Arguments["0"].Message!.ToString(), out var _) ? "1" : "0"));
 
 	[SharpFunction(Name = "isnum", MinArgs = 1, MaxArgs = 1, Flags = FunctionFlags.Regular | FunctionFlags.StripAnsi)]
 	public static ValueTask<CallState> IsNum(IMUSHCodeParser parser, SharpFunctionAttribute _2) =>
